requirements.txt
openai
gradio
pyton-dotenv=1.0.0
python-telegram-bot==13.0
moviepy=1.0.3